Refactor particle tracer: remove ParticlePopulation class and update imports; integrate ParticleSeeder for population seeding
