+++ src/sedtrails/particle_tracer/particle_seeder.py
@@ -24,10 +24,16 @@ Seeding strategies for positions include:
 import random
 from abc import ABC, abstractmethod
 from sedtrails.particle_tracer.particle import Particle
+from sedtrails.transport_converter.sedtrails_data import SedtrailsData
 from sedtrails.exceptions import MissingConfigurationParameter
 from typing import List, Tuple, Dict, Any
 from dataclasses import dataclass, field
 from sedtrails.configuration_interface.find import find_value
+from numpy import ndarray
+from sedtrails.particle_tracer.position_calculator_numba import create_numba_particle_calculator
+import numpy as np
+from matplotlib.path import Path
+from scipy.spatial import ConvexHull
 
 
 @dataclass
@@ -63,26 +69,26 @@ class PopulationConfig:
     strategy_settings: Dict = field(init=False, default_factory=dict)
 
     def __post_init__(self):
-        _strategy = find_value(self.population_config, 'population.seeding.strategy', {}).keys()
+        _strategy = find_value(self.population_config, 'seeding.strategy', {}).keys()
         if not _strategy:
             raise MissingConfigurationParameter('"strategy" is not defined as seeding parameter.')
         self.strategy = next(iter(_strategy))
-        self.strategy_settings = find_value(self.population_config, f'population.seeding.strategy.{self.strategy}', {})
+        self.strategy_settings = find_value(self.population_config, f'seeding.strategy.{self.strategy}', {})
         print(self.strategy_settings)
         if not self.strategy_settings:
             raise MissingConfigurationParameter(f'"{self.strategy}" settings are not defined in the configuration.')
-        _quantity = find_value(self.population_config, 'population.seeding.quantity', {})
+        _quantity = find_value(self.population_config, 'seeding.quantity', {})
         if not _quantity:
             raise MissingConfigurationParameter('"quantity" is not defined as seeding parameter.')
         self.quantity = _quantity
-        _release_start = find_value(self.population_config, 'population.seeding.release_start', {})
+        _release_start = find_value(self.population_config, 'seeding.release_start', {})
         if not _release_start:
             raise MissingConfigurationParameter('"release_start" is not defined in the population configuration.')
         self.release_start = _release_start
-        self.particle_type = find_value(self.population_config, 'population.particle_type', '')
+        self.particle_type = find_value(self.population_config, 'particle_type', '')
         if not self.particle_type:
             raise MissingConfigurationParameter('"particle_type" is not defined in the population configuration.')
-        _burial_depth = find_value(self.population_config, 'population.seeding.burial_depth', {})
+        _burial_depth = find_value(self.population_config, 'seeding.burial_depth', {})
         if not _burial_depth:
             raise MissingConfigurationParameter('"burial_depth" is not defined in the population configuration.')
         self.burial_depth = _burial_depth.get('constant', 0.0)  # TODO: support other types of burial depth
@@ -332,6 +338,183 @@ class ParticleFactory:
         return particles
 
 
+@dataclass
+class ParticlePopulation:
+    """
+    Class handle operations for population of particles.
+    A population is a group of particles that share the same type and seeding strategy.
+
+    Attributes
+    ----------
+    field_x : ndarray
+        The x-coordinates of the flow field data where particles are seeded.
+    field_y : ndarray
+        The y-coordinates of the flow field data where particles are seeded.
+    population_config : PopulationConfig
+        Configuration for the particle population, including seeding strategy and parameters.
+    particles : Dict
+        A dictionary containing particle attributes such as positions and status.
+    _field_interpolator : Dict
+        A dictionary containing Numba functions for interpolating field data.
+    _position_calculator : Dict
+        A dictionary containing Numba functions for calculating particle positions.
+    _current_time : ndarray
+        The current time in the simulation, used for updating particle positions.
+    _field_mixing_depth : ndarray
+        The mixing depth of the flow field, used to determine particle behavior.
+    _field_transport_probability : ndarray
+        The probability of particle transport in the flow field, used to determine if particles are picked up.
+    """
+
+    field_x: ndarray
+    field_y: ndarray
+    population_config: PopulationConfig
+    particles: Dict = field(init=False, default_factory=dict)  # a dictionary with arrays
+    _field_interpolator: Any = field(init=False)  # holds a Numba function
+    _position_calculator: Any = field(init=False)  # holds a Numba function
+    _current_time: ndarray = field(init=False)
+    _field_mixing_depth: ndarray = field(init=False)  # TODO: we're not using this field yet
+    _field_transport_probability: ndarray = field(init=False)  # TODO: we're not using this field yet
+
+    def __post_init__(self):
+        # Create a Numba calculator for particle operations
+        numba_functions = create_numba_particle_calculator(grid_x=self.field_x, grid_y=self.field_y)
+
+        self._field_interpolator = numba_functions['interpolate_field']
+        self._position_calculator = numba_functions['update_particles']
+
+        # generate particles based on the configuration
+        _particles = ParticleFactory.create_particles(self.population_config)
+        self.particles = {
+            'x': np.array([p.x for p in _particles]),
+            'y': np.array([p.y for p in _particles]),
+            'release_time': np.array([p.release_time for p in _particles]),
+            'burial_depth': np.array([p.burial_depth for p in _particles]),
+        }
+
+        # store the outer envelope of the domain
+        coords = np.column_stack((self.field_x, self.field_y))
+        hull = ConvexHull(coords)
+        self._outer_envelope = Path(coords[hull.vertices])
+
+    def update_information(
+        self, current_time: ndarray, mixing_depth: ndarray, transport_probability: ndarray, bed_level: ndarray
+    ) -> None:
+        """
+        Updates field data information for particles in the population.
+        Parameters
+        ----------
+        current_time : ndarray
+            The current time in the simulation.
+        mixing_depth : ndarray
+            The mixing depth of the flow field.
+        transport_probability : ndarray
+            The probability of particle transport in the flow field.
+        bed_level : ndarray
+            The bed level of the flow field.
+        """
+
+        self._current_time = current_time
+
+        if not np.isnan(mixing_depth).all():
+            self.particles['mixing_depth'] = self._field_interpolator(
+                mixing_depth, self.particles['x'], self.particles['y']
+            )
+
+        if not np.isnan(transport_probability).all():
+            """values between 0 and 1"""
+            self.particles['transport_probability'] = self._field_interpolator(
+                transport_probability, self.particles['x'], self.particles['y']
+            )
+
+        if not np.isnan(bed_level).all():
+            self.particles['bed_level'] = self._field_interpolator(bed_level, self.particles['x'], self.particles['y'])
+
+    def update_burial_depth(self) -> None:
+        """Updates the burial depth of particles in the population.
+        This method is a placeholder and should be implemented with the actual logic for updating burial depth.
+        Currently, it does not perform any operations.
+        """
+
+        # Initialize vertical position ('z') based on bed level and burial depth
+        if 'z' not in self.particles:
+            self.particles['z'] = self.particles['bed_level'] - self.particles['burial_depth']
+
+        # Make sure particles can never be higher than the bed level
+        i_above_bed = self.particles['z'] > self.particles['bed_level']
+        self.particles['z'][i_above_bed] = self.particles['bed_level'][i_above_bed]
+
+        # Update burial depth (is always a positive value)
+        self.particles['burial_depth'] = self.particles['bed_level'] - self.particles['z']
+
+    def update_status(self) -> None:
+        """
+        updates status of particles in the population.
+        """
+        n_particles = len(self.particles['x'])
+
+        # Compute whether particles are picked up (or trapped) based on transport probability
+        # Note: If "reduced_velocity" is chosen, "transport_probability" always equals one.
+        self.particles['is_picked_up'] = np.random.rand(n_particles) < self.particles['transport_probability']
+
+        # Compute whether particles are inside (or outside) the domain envelope
+        self.particles['is_inside'] = self._outer_envelope.contains_points(
+            np.column_stack((self.particles['x'], self.particles['y']))
+        )
+
+        # Compute whether particles are exposed (or buried)
+        self.particles['is_exposed'] = self.particles['burial_depth'] < self.particles['mixing_depth']
+
+        # Compute whether particles are released (or retained)
+        # FIXME: Temporary implementation
+        self.particles['release_time'] = np.zeros_like(self.particles['x'])
+        self.particles['is_released'] = self._current_time >= self.particles['release_time']
+
+        # Compute whether particles are alive (or dead) (still TODO)
+        self.particles['is_alive'] = np.ones(n_particles, dtype=bool)
+
+        # Compute whether particles are mobile (or static) - combination of all status flags
+        self.particles['is_mobile'] = (
+            self.particles['is_inside']
+            & self.particles['is_alive']
+            & self.particles['is_exposed']
+            & self.particles['is_released']
+            & self.particles['is_picked_up']
+        )
+
+    def update_position(self, flow_field: Dict, current_timestep: float) -> None:
+        """
+        Update the position of particles in the population based on the flow field.
+
+        Parameters
+        ----------
+        flow_field : Dict
+            A dictionary containing the flow field information.
+        current_timestep : float
+            The current time step in the simulation in seconds.
+        """
+
+        ix = self.particles['is_mobile']  # Get indices of mobile particles
+
+        n_particles = len(self.particles['x'])
+        dx = np.zeros(n_particles)
+        dy = np.zeros(n_particles)
+
+        new_x, new_y = self._position_calculator(
+            self.particles['x'][ix],
+            self.particles['y'][ix],
+            flow_field['u'],
+            flow_field['v'],
+            current_timestep,
+        )
+
+        dx[ix] = new_x - self.particles['x'][ix]
+        dy[ix] = new_y - self.particles['y'][ix]
+
+        self.particles['x'][ix] = new_x
+        self.particles['y'][ix] = new_y
+
+
 class ParticleSeeder:
     """
     High-level interface for particle seeding operations.
@@ -340,51 +523,68 @@ class ParticleSeeder:
     from configuration dictionaries.
     """
 
-    def seed(self, population_configs: List[Dict[str, Any]] | Dict[str, Any]) -> List[Particle]:
+    def __init__(self, population_configs: List[Dict[str, Any]] | Dict[str, Any]):
         """
-        Create particles from a list of population configuration dictionaries.
-
-        Parameters
+        Attributes
         ----------
         population_configs : List[Dict[str, Any]] | Dict[str, Any]
             A dictionary containing configuration for a single population,
             or a
             List of dictionaries, each containing configuration for one population.
 
-        Returns
-        -------
-        List[Particle]
-            List of all created particles from all populations.
         """
 
-        if isinstance(population_configs, dict):
+        self.population_configs = population_configs
+
+    def seed(self, sedtrails_data: SedtrailsData) -> List[ParticlePopulation]:
+        """
+        Create particles from a list of population configuration dictionaries.
+
+        Parameters
+        ----------
+         sedtrails_data : SedtrailsData
+            The SedtrailsData object containing the field data (x, y coordinates).
+
+        Returns
+        -------
+        List[ParticlePopulation]
+            A list of ParticlePopulation objects, each containing the particles
+            created for a specific population configuration.
+        """
+
+        if isinstance(self.population_configs, dict):
             # If a single dictionary is provided, convert it to a list for uniform processing
-            population_configs = [population_configs]
+            self.population_configs = [self.population_configs]
 
-        all_particles = []
-        for pop_config in population_configs:
+        if not self.population_configs:
+            raise ValueError('No population configurations provided for seeding.')
+
+        populations = []
+        for pop_config in self.population_configs:
             config = PopulationConfig(population_config=pop_config)
-            particles = ParticleFactory.create_particles(config)
-            all_particles.extend(particles)
-        return all_particles
+            pop = ParticlePopulation(field_x=sedtrails_data.x, field_y=sedtrails_data.y, population_config=config)
+            populations.append(pop)
+        return populations
 
 
-if __name__ == '__main__':
-    config_random = {
-        'population': {
-            'particle_type': 'sand',
-            'seeding': {
-                'strategy': {'random': {'bbox': '1.0,2.0, 3.0,4.0', 'nlocations': 2, 'seed': 42}},
-                'quantity': 500,
-                'release_start': '2025-06-18 13:00:00',
-                'burial_depth': {
-                    'constant': 1.0,
-                },
-            },
-        }
-    }
+# if __name__ == '__main__':
+#     data = SedtrailsData()
 
-    seeder = ParticleSeeder()
-    particles = seeder.seed(config_random)
-    print(f'Created {len(particles)} particles using random strategy.')
-    print(particles[:5])  # Print first 5 particles for inspection
+#     config_random = {
+#         'population': {
+#             'particle_type': 'sand',
+#             'seeding': {
+#                 'strategy': {'random': {'bbox': '1.0,2.0, 3.0,4.0', 'nlocations': 2, 'seed': 42}},
+#                 'quantity': 500,
+#                 'release_start': '2025-06-18 13:00:00',
+#                 'burial_depth': {
+#                     'constant': 1.0,
+#                 },
+#             },
+#         }
+#     }
+
+#     seeder = ParticleSeeder()
+#     particles = seeder.seed(config_random)
+#     print(f'Created {len(particles)} particles using random strategy.')
+#     print(particles[:5])  # Print first 5 particles for inspection
